SearchBar changes query instead of emmiting
